requirements.txt
feedparser
python-telegram-bot
python-dateutil
FilterHTML
psycopg2
bs4